Move loan period and accessories to radio orders and remove UI for it
Structure with accessories and loan period per radio was too detailed.
This moves those to radio orders instead which should make things
clearer. The UI is removed to make place for new and simpler solutions.

Migrations are done by moving the current tables to an archive and
recreating the tables with new structure in place to preserve any data.